requirements.txt
asgiref==3.8.1
Django==5.1.1
sqlparse==0.5.1
typing_extensions==4.12.2
django-haystack==3.3.0
elasticsearch==7.17.9
python-dotenv==1.0.1